Fixed problem with initial SVG favicon
In processes that didn't load the site's javascript and preferred SVG favicons (e.g. Firefox home screen), the displayed favicon for Weird Waves would be the bare-bones invisible one. This fixes that by adding a default SVG favicon that looks like the default .ico favicon, but has 0 opacity for 1 second (long enough for the script to seamlessly load in the themed favicon when at the actual site, before the default appears).